requirements.txt
tensorflow==2.3.1
joblib==0.14.1
numpy==1.16.4
pandas==0.25.3
scipy==1.4.1
python_dateutil==2.8.1
scikit_learn==0.23.2
